refactor: remove message file
