+++ packages/engine/src/vaults/impl/evm/sdk/index.ts
@@ -1,3 +1,2 @@
 export * from './ethers';
 export * from './geth';
-export * from './message';
